backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.37.0
databases==0.9.0
aiomysql==0.2.0
PyMySQL==1.1.2
SQLAlchemy==2.0.43
python-dotenv==1.1.1
passlib[argon2]==1.7.4
google-genai
python-multipart
python-doctr
boto3==1.35.77